+++ src/index.ts
@@ -18,6 +18,7 @@ export async function runner(
 			runId: process.env.AGENTUITY_CLOUD_RUN_ID,
 			port: parseInt(process.env.PORT ?? '3000'),
 			devmode: process.env.AGENTUITY_SDK_DEV_MODE === 'true',
+			cliVersion: process.env.AGENTUITY_CLI_VERSION,
 			environment:
 				process.env.AGENTUITY_ENVIRONMENT ??
 				process.env.NODE_ENV ??