+++ src/otel/index.ts
@@ -14,14 +14,21 @@ import {
 	ATTR_SERVICE_VERSION,
 } from '@opentelemetry/semantic-conventions';
 import opentelemetry, { type Tracer } from '@opentelemetry/api';
+import {
+	W3CTraceContextPropagator,
+	W3CBaggagePropagator,
+	CompositePropagator,
+} from '@opentelemetry/core';
 import * as LogsAPI from '@opentelemetry/api-logs';
 import {
 	LoggerProvider,
 	SimpleLogRecordProcessor,
 } from '@opentelemetry/sdk-logs';
 import { OTLPLogExporter } from '@opentelemetry/exporter-logs-otlp-http';
-import { createLogger } from './logger';
+import { CompressionAlgorithm } from '@opentelemetry/otlp-exporter-base';
+import { createLogger, patchConsole } from './logger';
 import { ConsoleLogRecordExporter } from './console';
+import { instrumentFetch } from './fetch';
 
 interface OtelConfig {
 	url?: string;
@@ -34,6 +41,7 @@ interface OtelConfig {
 	runId?: string;
 	environment?: string;
 	sdkVersion?: string;
+	cliVersion?: string;
 }
 
 interface OtelResponse {
@@ -50,6 +58,7 @@ export function registerOtel(config: OtelConfig): OtelResponse {
 		bearerToken,
 		environment = 'development',
 		sdkVersion,
+		cliVersion,
 		orgId,
 		projectId,
 		deploymentId,
@@ -72,13 +81,19 @@ export function registerOtel(config: OtelConfig): OtelResponse {
 		'@agentuity/runId': runId ?? 'unknown',
 		'@agentuity/env': environment,
 		'@agentuity/sdkVersion': sdkVersion ?? 'unknown',
+		'@agentuity/cliVersion': cliVersion ?? 'unknown',
 	});
 
 	let otlpLogExporter: OTLPLogExporter | undefined;
 	let logRecordProcessor: SimpleLogRecordProcessor | undefined;
 
 	if (url) {
-		otlpLogExporter = new OTLPLogExporter({ url: `${url}/v1/logs`, headers });
+		otlpLogExporter = new OTLPLogExporter({
+			url: `${url}/v1/logs`,
+			headers,
+			compression: CompressionAlgorithm.GZIP,
+			timeoutMillis: 10_000,
+		});
 		logRecordProcessor = new SimpleLogRecordProcessor(otlpLogExporter);
 	} else {
 		logRecordProcessor = new SimpleLogRecordProcessor(
@@ -93,6 +108,15 @@ export function registerOtel(config: OtelConfig): OtelResponse {
 	LogsAPI.logs.setGlobalLoggerProvider(loggerProvider);
 
 	const logger = createLogger(!!url);
+
+	// must do this after we have created the logger
+	patchConsole({
+		'@agentuity/orgId': orgId ?? 'unknown',
+		'@agentuity/projectId': projectId ?? 'unknown',
+		'@agentuity/deploymentId': deploymentId ?? 'unknown',
+		'@agentuity/runId': runId ?? 'unknown',
+		'@agentuity/env': environment,
+	});
 
 	const traceExporter = url
 		? new OTLPTraceExporter({
@@ -139,12 +163,19 @@ export function registerOtel(config: OtelConfig): OtelResponse {
 	let instrumentationSDK: NodeSDK | undefined;
 
 	if (url) {
+		instrumentFetch();
 		instrumentationSDK = new NodeSDK({
 			logRecordProcessor,
 			traceExporter,
 			metricReader: sdkMetricReader,
 			instrumentations: [getNodeAutoInstrumentations()],
 			resource,
+			textMapPropagator: new CompositePropagator({
+				propagators: [
+					new W3CTraceContextPropagator(),
+					new W3CBaggagePropagator(),
+				],
+			}),
 		});
 		instrumentationSDK.start();
 		hostMetrics?.start();
@@ -157,8 +188,19 @@ export function registerOtel(config: OtelConfig): OtelResponse {
 		if (running) {
 			running = false;
 			logger.debug('shutting down OpenTelemetry');
-			await instrumentationSDK?.shutdown();
-			await otlpLogExporter?.shutdown();
+			await otlpLogExporter
+				?.forceFlush()
+				.catch((e) =>
+					logger.warn('error in forceFlush of otel exporter. %s', e)
+				);
+			await otlpLogExporter
+				?.shutdown()
+				.catch((e) => logger.warn('error in shutdown of otel exporter. %s', e));
+			await instrumentationSDK
+				?.shutdown()
+				.catch((e) =>
+					logger.warn('error in shutdown of otel instrumentation. %s', e)
+				);
 			logger.debug('shut down OpenTelemetry');
 		}
 	};