+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # @agentuity/sdk
 
+## 0.0.64
+
+### Patch Changes
+
+- Add fetch instrumentation
+- Add otel trace propagation
+- Monkey patch console to redirect into otel logger
+- More improvements on bundling JS
+- Handle direct return from agent run in case you don't return correct signature
+- Set the span status on completion for agent run
+- Add cli version to trace context
+- More work on agent-to-agent remote comms
+
 ## 0.0.63
 
 ### Patch Changes